builder: move logger and add alias for production build
