+++ packages/builder/package.json
@@ -28,7 +28,7 @@
   },
   "homepage": "https://github.com/abuxvn/source/tree/main/packages/builder#readme",
   "dependencies": {
-    "@abux/cli-collapse": "workspace:^",
+    "@abux/logger": "workspace:^",
     "@abux/resolve": "workspace:^",
     "chalk": "^4",
     "commander": "^10.0.1",
