+++ packages/builder/src/cmd/options.ts
@@ -7,3 +7,6 @@ export const nodeEnv = new Option('--node-env <env>', 'Specify build environment
 
 export const path = new Option('--path <path>', 'Specify root path for compilation')
   .default(process.cwd())
+
+export const production = new Option('--production', 'Production build')
+  .default(false)
