+++ packages/builder/src/lib/entries.ts
@@ -1,7 +1,9 @@
 import glob from 'fast-glob'
+import { logInfo, bold, italic, log } from '@abux/logger'
 import type {
   ITargetedExpandedEntries,
-  IPathResolver
+  IPathResolver,
+  IWebpackConfig
 } from '../interfaces'
 import { resolve } from './paths'
 
@@ -29,4 +31,13 @@ export const expandEntries = async (
     },
     {}
   )
+}
+
+export const logEntries = (configs: IWebpackConfig[]) => {
+  logInfo(bold.cyanBright('Building entries:'))
+
+  configs.forEach(({ name, target, entry }) => {
+    log(`   ${name} (${italic(target)}):`)
+    Object.keys(entry).forEach(entryName => { log(`     ${entryName}`) })
+  })
 }
