+++ packages/builder/src/plugins/DtsPlugin/index.js
@@ -1,1 +1,1 @@
-(()=>{"use strict";var e={559:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.badge=t.colorIndex=t.color=t.logEntries=t.logSuccess=t.logError=t.logWarn=t.logProgress=t.logInfo=t.log=t.bold=void 0;const o=r(22);var n=r(22);Object.defineProperty(t,"bold",{enumerable:!0,get:function(){return n.bold}}),t.log=console.log.bind(console),t.logInfo=(...e)=>{console.info(o.bold.blueBright("ℹ"),...e)},t.logProgress=(...e)=>{console.info((0,o.gray)("➤"),...e)},t.logWarn=(...e)=>{console.info((0,o.red)("△"),...e)},t.logError=(...e)=>{e.forEach((e=>{console.error(o.bold.underline.redBright("✘"),e)}))},t.logSuccess=(...e)=>{console.info(o.bold.greenBright("✔"),...e)},t.logEntries=e=>{(0,t.logInfo)(o.bold.cyanBright("Building entries:")),e.forEach((({name:e,target:r,entry:n})=>{(0,t.log)(`   ${e} (${(0,o.italic)(r)}):`),Object.keys(n).forEach((e=>{(0,t.log)(`     ${e}`)}))}))};const i={green:o.green,yellow:o.yellow,blue:o.blue,magenta:o.magenta,cyan:o.cyan,greenBright:o.greenBright,yellowBright:o.yellowBright,blueBright:o.blueBright,magentaBright:o.magentaBright,cyanBright:o.cyanBright},l=Object.keys(i),a={green:o.bgGreen,yellow:o.bgYellow,blue:o.bgBlue,magenta:o.bgMagenta,cyan:o.bgCyan,greenBright:o.bgGreenBright,yellowBright:o.bgYellowBright,blueBright:o.bgBlueBright,magentaBright:o.bgMagentaBright,cyanBright:o.bgCyanBright,red:o.bgRed,redBright:o.bgRedBright};t.color=(e,t="blue")=>t?i[t](e):e,t.colorIndex=(e,r=0)=>{const o=l[r%l.length];return(0,t.color)(e,o)},t.badge=(e,t="blueBright")=>t?a[t](` ${(0,o.bold)(e)} `):e},584:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.getLocalPackagePath=t.removeExt=t.merge=t.resolve=t.resolver=t.getName=t.getDir=t.normalize=t.PathResolver=void 0;const o=r(17);class n{constructor(e){this.rootPath=(0,t.normalize)((0,o.resolve)(e))}relative(e){return(0,t.normalize)((0,o.relative)(this.rootPath,(0,t.normalize)(e)))}relativeList(e){return e.map((e=>this.relative(e)))}includes(e){return 0===(0,t.normalize)(e).indexOf(this.rootPath)}resolve(...e){return(0,t.normalize)((0,o.resolve)(this.rootPath,...e.filter(Boolean).map((e=>e.replace(/^\/+/,"")))))}resolveList(e){return e.map((e=>this.resolve(e)))}dir(){return(0,t.resolver)((0,t.getDir)(this.rootPath))}res(...e){return(0,t.resolver)(this.resolve(...e))}}t.PathResolver=n,t.normalize=e=>(null==e?void 0:e.replace(/\\/g,"/"))||"",t.getDir=e=>(0,t.normalize)(e).replace(/\/[^/]+\/?$/,""),t.getName=e=>(0,o.basename)((0,t.normalize)(e)),t.resolver=e=>new n(e),t.resolve=e=>(0,t.normalize)((0,o.resolve)(e)),t.merge=(...e)=>(0,t.normalize)((0,o.join)(...e)),t.removeExt=e=>null==e?void 0:e.replace(/\.([^/]+)$/,"");const i=/([^/.]+\/[^/.]+)/;t.getLocalPackagePath=e=>{if(e.includes("node_modules")||e.includes(".yarn"))return"";const t=e.match(i);return t?t[1]:""}},285:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.mkdirp=t.createWriteStream=t.writeJSON=t.readFile=t.readJSON=t.pathExists=t.copy=t.stat=void 0;var o=r(470);Object.defineProperty(t,"stat",{enumerable:!0,get:function(){return o.stat}}),Object.defineProperty(t,"copy",{enumerable:!0,get:function(){return o.copy}}),Object.defineProperty(t,"pathExists",{enumerable:!0,get:function(){return o.pathExists}}),Object.defineProperty(t,"readJSON",{enumerable:!0,get:function(){return o.readJSON}}),Object.defineProperty(t,"readFile",{enumerable:!0,get:function(){return o.readFile}}),Object.defineProperty(t,"writeJSON",{enumerable:!0,get:function(){return o.writeJSON}}),Object.defineProperty(t,"createWriteStream",{enumerable:!0,get:function(){return o.createWriteStream}}),Object.defineProperty(t,"mkdirp",{enumerable:!0,get:function(){return o.mkdirp}})},163:function(e,t,r){var o=this&&this.__awaiter||function(e,t,r,o){return new(r||(r=Promise))((function(n,i){function l(e){try{s(o.next(e))}catch(e){i(e)}}function a(e){try{s(o.throw(e))}catch(e){i(e)}}function s(e){var t;e.done?n(e.value):(t=e.value,t instanceof r?t:new r((function(e){e(t)}))).then(l,a)}s((o=o.apply(e,t||[])).next())}))};Object.defineProperty(t,"__esModule",{value:!0}),t.DtsPlugin=void 0;const n=r(903),i=r(285),l=r(559),a=r(584);let s=0;t.DtsPlugin=class{constructor(e){this.path=(0,a.resolver)(e)}apply(e){let t=[];e.hooks.beforeCompile.tapPromise("[dts] start collecting built modules",(()=>o(this,void 0,void 0,(function*(){t=[]})))),e.hooks.compilation.tap("[dts] setup compilation",(e=>{e.hooks.succeedModule.tap("[dts] collect built module",(e=>{if("NormalModule"!==e.constructor.name)return;const r=this.path.relative(e.context||""),o=(0,a.getLocalPackagePath)(r);o&&!t.includes(o)&&t.push(o)}))})),e.hooks.afterCompile.tap("[dts] generate definitions",(()=>{Promise.all(t.map((e=>o(this,void 0,void 0,(function*(){const t=s+=1;try{const r=yield(0,i.readJSON)(this.path.resolve(e,"package.json")),o=r.types,s=r.name,c=this.path.resolve(e),g=this.path.resolve(e,o),u=r.main||"index",d=r.files||"";if(!o)return;let h=this.path.resolve(e,"tsconfig.json");if((yield(0,i.pathExists)(h))||(h=this.path.resolve("tsconfig.json")),!(yield(0,i.pathExists)(h)))return void(0,l.logWarn)(this.log(t,s,"generation ignored, required tsconfig"));const p=new n.Dts,m=d.map((t=>(0,a.resolver)(c).relative(this.path.resolve(e,t))));p.on("log",(e=>{(0,l.logProgress)(this.log(t,e))})),(0,l.logInfo)(this.log(t,s,"generation started")),yield p.generate({projectPath:h,name:s,inputDir:c,outputPath:g,main:(0,a.removeExt)(u.replace(/^(\.\/?)+/,"")),filePatterns:m}),(0,l.logSuccess)(this.log(t,s,"declaration at",o))}catch(e){(0,l.logError)(this.log(t,e.message))}})))))}))}log(e,...t){let r=t.join(" ");return 0!==r.indexOf("[dts")&&(r=`[dts] ${r}`),r.replace(/^\[(dtsw?)\]/,((t,r)=>(0,l.colorIndex)(`[${r} ${e}]`,e)))}}},22:e=>{e.exports=require("chalk")},470:e=>{e.exports=require("fs-extra")},903:e=>{e.exports=require("@abux/builder/src/lib/dts/index.js")},17:e=>{e.exports=require("path")}},t={},r=function r(o){var n=t[o];if(void 0!==n)return n.exports;var i=t[o]={exports:{}};return e[o].call(i.exports,i,i.exports,r),i.exports}(163);module.exports=r})();
+(()=>{"use strict";var e={584:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.getLocalPackagePath=t.removeExt=t.merge=t.resolve=t.resolver=t.getName=t.getDir=t.normalize=t.PathResolver=void 0;const o=r(17);class i{constructor(e){this.rootPath=(0,t.normalize)((0,o.resolve)(e))}relative(e){return(0,t.normalize)((0,o.relative)(this.rootPath,(0,t.normalize)(e)))}relativeList(e){return e.map((e=>this.relative(e)))}includes(e){return 0===(0,t.normalize)(e).indexOf(this.rootPath)}resolve(...e){return(0,t.normalize)((0,o.resolve)(this.rootPath,...e.filter(Boolean).map((e=>e.replace(/^\/+/,"")))))}resolveList(e){return e.map((e=>this.resolve(e)))}dir(){return(0,t.resolver)((0,t.getDir)(this.rootPath))}res(...e){return(0,t.resolver)(this.resolve(...e))}}t.PathResolver=i,t.normalize=e=>(null==e?void 0:e.replace(/\\/g,"/"))||"",t.getDir=e=>(0,t.normalize)(e).replace(/\/[^/]+\/?$/,""),t.getName=e=>(0,o.basename)((0,t.normalize)(e)),t.resolver=e=>new i(e),t.resolve=e=>(0,t.normalize)((0,o.resolve)(e)),t.merge=(...e)=>(0,t.normalize)((0,o.join)(...e)),t.removeExt=e=>null==e?void 0:e.replace(/\.([^/]+)$/,"");const n=/([^/.]+\/[^/.]+)/;t.getLocalPackagePath=e=>{if(e.includes("node_modules")||e.includes(".yarn"))return"";const t=e.match(n);return t?t[1]:""}},285:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.mkdirp=t.createWriteStream=t.writeJSON=t.readFile=t.readJSON=t.pathExists=t.copy=t.stat=void 0;var o=r(470);Object.defineProperty(t,"stat",{enumerable:!0,get:function(){return o.stat}}),Object.defineProperty(t,"copy",{enumerable:!0,get:function(){return o.copy}}),Object.defineProperty(t,"pathExists",{enumerable:!0,get:function(){return o.pathExists}}),Object.defineProperty(t,"readJSON",{enumerable:!0,get:function(){return o.readJSON}}),Object.defineProperty(t,"readFile",{enumerable:!0,get:function(){return o.readFile}}),Object.defineProperty(t,"writeJSON",{enumerable:!0,get:function(){return o.writeJSON}}),Object.defineProperty(t,"createWriteStream",{enumerable:!0,get:function(){return o.createWriteStream}}),Object.defineProperty(t,"mkdirp",{enumerable:!0,get:function(){return o.mkdirp}})},163:function(e,t,r){var o=this&&this.__awaiter||function(e,t,r,o){return new(r||(r=Promise))((function(i,n){function a(e){try{l(o.next(e))}catch(e){n(e)}}function s(e){try{l(o.throw(e))}catch(e){n(e)}}function l(e){var t;e.done?i(e.value):(t=e.value,t instanceof r?t:new r((function(e){e(t)}))).then(a,s)}l((o=o.apply(e,t||[])).next())}))};Object.defineProperty(t,"__esModule",{value:!0}),t.DtsPlugin=void 0;const i=r(903),n=r(285),a=r(893),s=r(584);let l=0;t.DtsPlugin=class{constructor(e){this.path=(0,s.resolver)(e)}apply(e){let t=[];e.hooks.beforeCompile.tapPromise("[dts] start collecting built modules",(()=>o(this,void 0,void 0,(function*(){t=[]})))),e.hooks.compilation.tap("[dts] setup compilation",(e=>{e.hooks.succeedModule.tap("[dts] collect built module",(e=>{if("NormalModule"!==e.constructor.name)return;const r=this.path.relative(e.context||""),o=(0,s.getLocalPackagePath)(r);o&&!t.includes(o)&&t.push(o)}))})),e.hooks.afterCompile.tap("[dts] generate definitions",(()=>{Promise.all(t.map((e=>o(this,void 0,void 0,(function*(){const t=l+=1;try{const r=yield(0,n.readJSON)(this.path.resolve(e,"package.json")),o=r.types,l=r.name,c=this.path.resolve(e),u=this.path.resolve(e,o),d=r.main||"index",p=r.files||"";if(!o)return;let h=this.path.resolve(e,"tsconfig.json");if((yield(0,n.pathExists)(h))||(h=this.path.resolve("tsconfig.json")),!(yield(0,n.pathExists)(h)))return void(0,a.logWarn)(this.log(t,l,"generation ignored, required tsconfig"));const m=new i.Dts,v=p.map((t=>(0,s.resolver)(c).relative(this.path.resolve(e,t))));m.on("log",(e=>{(0,a.logProgress)(this.log(t,e))})),(0,a.logInfo)(this.log(t,l,"generation started")),yield m.generate({projectPath:h,name:l,inputDir:c,outputPath:u,main:(0,s.removeExt)(d.replace(/^(\.\/?)+/,"")),filePatterns:v}),(0,a.logSuccess)(this.log(t,l,"declaration at",o))}catch(e){(0,a.logError)(this.log(t,e.message))}})))))}))}log(e,...t){let r=t.join(" ");return 0!==r.indexOf("[dts")&&(r=`[dts] ${r}`),r.replace(/^\[(dtsw?)\]/,((t,r)=>(0,a.color)(`[${r} ${e}]`,e)))}}},893:e=>{e.exports=require("@abux/logger")},470:e=>{e.exports=require("fs-extra")},903:e=>{e.exports=require("@abux/builder/src/lib/dts/index.js")},17:e=>{e.exports=require("path")}},t={},r=function r(o){var i=t[o];if(void 0!==i)return i.exports;var n=t[o]={exports:{}};return e[o].call(n.exports,n,n.exports,r),n.exports}(163);module.exports=r})();
