+++ packages/logger/README.md
@@ -1,0 +1,8 @@
+@abux/logger
+=====
+
+![build][badge-build]
+
+An utility logger for CLI and web, to be reused in our projects
+
+[badge-build]: https://github.com/abuxvn/source/actions/workflows/build.yaml/badge.svg
